+++ CaseStudies/Mandelbrot/MandelbrotBase.h
@@ -27,7 +27,7 @@ protected:
     size_t m_clientHeight;
 
 protected:
-    virtual void drawPixel(HDC hdc, int x, int y, COLORREF color) const = 0;
+    virtual void drawPixel(HDC hdc, size_t x, size_t  y, COLORREF color) const = 0;
 
 public:
     // getter/setter
