+++ CaseStudies/Mandelbrot/MandelbrotBasic.cpp
@@ -15,9 +15,9 @@ extern MandelbrotPalette g_palette;
 // c'tor(s)
 MandelbrotBasic::MandelbrotBasic() {}
 
-void MandelbrotBasic::drawPixel(HDC hdc, int x, int y, COLORREF color) const
+void MandelbrotBasic::drawPixel(HDC hdc, size_t x, size_t y, COLORREF color) const
 {
-    ::SetPixelV(hdc, x, y, color);
+    ::SetPixelV(hdc, (int) x, (int) y, color);
 }
 
 // =====================================================================================
