+++ CaseStudies/Mandelbrot/MandelbrotRectanglesParallelBlocking.cpp
@@ -78,11 +78,11 @@ void MandelbrotRectanglesParallelBlocking::paintRectanglesAsync (HDC hDC) const 
 }
 
 
-void MandelbrotRectanglesParallelBlocking::drawPixel(HDC hdc, int x, int y, COLORREF color) const
+void MandelbrotRectanglesParallelBlocking::drawPixel(HDC hdc, size_t x, size_t y, COLORREF color) const
 {
     // RAII lock
     std::lock_guard<std::mutex> lock{ m_mutex };
-    ::SetPixelV(hdc, x, y, color);
+    ::SetPixelV(hdc, (int) x, (int) y, color);
 }
 
 // =====================================================================================
