+++ CaseStudies/Mandelbrot/MandelbrotRectanglesParallelBlockingUsingLatch.cpp
@@ -88,11 +88,11 @@ void MandelbrotRectanglesParallelBlockingUsingLatch::paintRectanglesAsyncWithLat
     }
 }
 
-void MandelbrotRectanglesParallelBlockingUsingLatch::drawPixel(HDC hdc, int x, int y, COLORREF color) const
+void MandelbrotRectanglesParallelBlockingUsingLatch::drawPixel(HDC hdc, size_t x, size_t y, COLORREF color) const
 {
     // RAII lock
     std::lock_guard<std::mutex> lock{ m_mutex };
-    ::SetPixelV(hdc, x, y, color);
+    ::SetPixelV(hdc, (int) x, (int) y, color);
 }
 
 // =====================================================================================
