+++ CaseStudies/Mandelbrot/MandelbrotRectanglesParallelNonBlockingClassic.h
@@ -44,7 +44,7 @@ private:
     size_t paintRectangle(HWND, HDC, struct Rectangle);
 
 private:
-    virtual void drawPixel(HDC, int x, int y, COLORREF) const override;
+    virtual void drawPixel(HDC, size_t x, size_t y, COLORREF) const override;
 };
 
 // =====================================================================================
